+++ spec/active_triples/rdf_source_spec.rb
@@ -1,0 +1,5 @@
+require 'spec_helper'
+
+describe ActiveTriples::RDFSource do
+  let(:dummy_source) { Class.new { include described_class } }
+end
